requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
google-genai>=1.0.0
Werkzeug==3.0.3
